Product calc

+++ database/migrations/2020_12_19_173411_create_taxes_table.php
@@ -17,7 +17,7 @@ class CreateTaxesTable extends Migration
             $table->id();
 
             $table->string('name');
-            $table->integer('tax_val');
+            $table->decimal('tax_val');
 
             $table->timestamps();
         });
